requirements2.txt
scapy
netifaces
elasticsearch
certifi
maxminddb-geolite2